fix: recall test and delete consolidation

+++ fresh_vamana_recall_test.cpp
@@ -3,6 +3,7 @@
 #include <chrono>
 #include <set>
 #include <map>
+#include <cassert>
 #include "FreshVamanaTestUtils.hpp"
 #include "FreshVamanaIndex.h"
 
@@ -11,11 +12,12 @@ int main() {
     // Test parameters (annotated paper values)
     const size_t NEIGHBOR_COUNT = 5, // 5
     SEARCH_LIST_SIZE = 30, // 75
-    N_TEST_POINTS = 100,
+    N_TEST_POINTS = 50,
             OUT_DEGREE_BOUND = 64; // 64
-    const double ALPHA = 1.2; // 1.2
+    const double ALPHA = 1.2, // 1.2
+    DELETE_ACCUMULATION_FACTOR = 0.04; // 0.1
 
-    FreshVamanaIndex index(ALPHA, OUT_DEGREE_BOUND);
+    FreshVamanaIndex index(ALPHA, OUT_DEGREE_BOUND, DELETE_ACCUMULATION_FACTOR);
 
     auto dataset = FreshVamanaTestUtils::loadDataset("siftsmall_base.csv");
 
@@ -31,16 +33,46 @@ int main() {
 
     auto nearestMap = FreshVamanaTestUtils::loadNearestGroundTruth("siftsmall_groundtruth.csv");
     auto queries = FreshVamanaTestUtils::loadDataset("siftsmall_query.csv");
+    // use only a subset of the queries for this test
+    auto rng = std::default_random_engine{};
+    std::shuffle(queries.begin(), queries.end(), rng);
+    queries.resize(N_TEST_POINTS);
+
+    // results file
+    std::ofstream outfile("results.csv");
+    if (!outfile.is_open()) {
+        std::cerr << "Error: Could not open file " << "results.csv" << std::endl;
+        return -1;
+    }
+
+    std::cout << std::setprecision(5);
+    outfile << std::setprecision(5);
+
+    double removeThreshold = 0.05 * dataset.size();
 
     // Test cycles of insertion / deletion
     size_t cycle = 0;
+    std::map<size_t, std::shared_ptr<GraphNode>> removed;
     while (cycle < 20) {
-        std::map<size_t, std::shared_ptr<GraphNode>> removed;
         // Insert and delete random points
         auto randomPoint = FreshVamanaTestUtils::pickRandomPoint(dataset, removed);
         index.deleteNode(randomPoint);
+        removed.insert({randomPoint->id, randomPoint});
 
-        size_t testCnt = 0;
+        // assert removal
+//        for (const auto &node : index.graph) {
+//            if (node->deleted) {
+//                if (index.deleteList.find(node) != index.deleteList.end()) {
+//                    continue;
+//                }
+//                if (removed.find(node->id) == removed.end()) {
+//                    // node should be in the graph
+//                    std::cout << node->id << std::endl;
+//                    std::cout << index.graph.size() << std::endl;
+//                }
+//            }
+//        }
+
         double avgRecall = 0;
         for (auto queryPoint: queries) {
             auto timedResult = FreshVamanaTestUtils::time_function<std::vector<std::shared_ptr<GraphNode>>>([&]() {
@@ -49,9 +81,8 @@ int main() {
 //            std::cout << "Search took " << timedResult.duration << " ms" << std::endl;
             // Verify search correctness
             auto trueNeighbors = nearestMap[queryPoint->id];
-//            trueNeighbors.resize(NEIGHBOR_COUNT);
             std::set<size_t> trueNeighborSet;
-            for (const auto &actualNeighbor : trueNeighbors) {
+            for (const auto &actualNeighbor: trueNeighbors) {
                 if (removed.find(actualNeighbor) == removed.end()) {
                     trueNeighborSet.insert(actualNeighbor);
                     if (trueNeighborSet.size() == NEIGHBOR_COUNT) {
@@ -65,11 +96,22 @@ int main() {
             for (size_t i = 0; i < NEIGHBOR_COUNT; ++i) {
                 size_t foundNeighbor = timedResult.result[i]->id;
                 auto foundNeighborNode = timedResult.result[i];
+
                 auto trueNeighborNode = index.getNode(trueNeighborVec[i]);
-                std::cout << i + 1 << ": " << "(TRUE) " << trueNeighborVec[i] << " with distance "
-                          << index.distance(trueNeighborNode, queryPoint) << " (FOUND) "
-                          << foundNeighbor << " with distance " << index.distance(foundNeighborNode, queryPoint)
-                          << std::endl;
+                // assert
+                assert(trueNeighborNode != nullptr);
+                if (trueNeighborNode == nullptr) {
+                    std::cout << trueNeighborVec[i] << std::endl;
+                    std::cout << (removed.find(trueNeighborVec[i]) == removed.end()) << std::endl;
+                    std::cout << index.graph.size() << std::endl;
+                    std::cout << index.deleteList.size() << std::endl;
+                }
+
+//                std::cout << i + 1 << ": " << "(TRUE) " << trueNeighborVec[i] << " with distance "
+//                          << index.distance(trueNeighborNode, queryPoint) << " (FOUND) "
+//                          << foundNeighbor << " with distance " << index.distance(foundNeighborNode, queryPoint)
+//                          << std::endl;
+
                 // Count for recall
                 if (trueNeighborSet.find(foundNeighbor) != trueNeighborSet.end()) {
                     positiveCount++;
@@ -77,28 +119,42 @@ int main() {
             }
             double recall = ((double) positiveCount / (double) NEIGHBOR_COUNT);
             avgRecall += recall;
-            std::cout << "recall@" << NEIGHBOR_COUNT << ": " << recall
-                      << std::endl;
-            testCnt++;
-            if (testCnt == N_TEST_POINTS) {
-                break;
-            }
+//            std::cout << "recall@" << NEIGHBOR_COUNT << ": " << recall
+//                      << std::endl;
         }
-        std::cout << "avg recall@" << NEIGHBOR_COUNT << ": " << (avgRecall / (double) N_TEST_POINTS)
+        double progress = (double) cycle + ((double) removed.size() / removeThreshold);
+        avgRecall = (avgRecall / (double) N_TEST_POINTS);
+        std::cout << "progress: " << progress << " avg recall@" << NEIGHBOR_COUNT << ": " << avgRecall
                   << std::endl;
 
-        if (removed.size() >= 0.05 * dataset.size()) {
-            std::cout << "Finished cycle " << cycle << " reinserting..." << std::endl;
+        // write to file
+
+        outfile << progress << "," << avgRecall << "\n";
+
+        if (removed.size() >= removeThreshold) {
+            std::cout << "Finished cycle " << cycle << " . Reinserting..." << std::endl;
             while (!removed.empty()) {
-                index.insert(removed.begin()->second);
+                auto removedNode  = removed.begin()->second;
+                // make sure node is not marked as deleted
+                removedNode->deleted = false;
+                if (index.deleteList.find(removedNode) == index.deleteList.end()) {
+                    index.insert(removedNode);
+                }
+//                else {
+//                    // assert
+//                    // dont reinsert deleted nodes (shouldnt happen with 0.04 delete factor)
+//                    std::cout << "assert delete factor" << std::endl;
+//                    std::cout << removedNode->id << std::endl;
+//                    std::cout << index.graph.size() << std::endl;
+//                    std::cout << index.deleteList.size() << std::endl;
+//                }
                 removed.erase(removed.begin());
             }
+            assert(index.graph.size() == dataset.size());
             cycle++;
-
         }
     }
-
-
+    outfile.close();
 
 
     return 0;
